requirements/dev.txt
-r common.txt

coverage>=4.3.4
coveralls>=1.1
flake8>=3.2.1
pytest>=3.0.6
pytest-django>=3.1.2
pytest-factoryboy>=1.3.0
pytest-faker>=2.0.0
pytest-mock>=1.5.0
tox>=2.5.0
